CanSat-Ground-station-UTEC-FORK/requirements.txt
numpy
PyQt5
PyQt5-sip
pyqtgraph
pyserial
ipyleaflet
folium
PyQtWebEngine
Pillow